path management improvements back to u64 File<Local> doesn't keep track of size anymore try using bindings instead of a single reference to a big object more monolithic approach for path dispatching, let BDrive struct handle the relativity of paths

+++ src/db/file.rs
@@ -11,11 +11,11 @@ pub struct Sync;
 pub struct RemoteFile {
     pub path: String,
     pub hash: String,
-    pub size: u32,
+    pub size: u64,
 }
 
 impl RemoteFile {
-    pub fn new(path: String, hash: String, size: u32) -> Self {
+    pub fn new(path: String, hash: String, size: u64) -> Self {
         Self { path, hash, size }
     }
